Move HSL parsing to `hsl` color space module
This encapsulates the color parsing and testing in the same module as
the rest of the functionality for the HSL color space.
